requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.0.5
djangorestframework==3.11.0
idna==2.8
pytz==2019.3
requests==2.22.0
urllib3==1.25.7
